fix crash in exporter
It arose because the web server is running in a thread, and so the
update() and format() methods could step on each other.  This is a
minimal solution, which protects the self.metrics dictionaries with a
lock.  A longer-term solution would be to switch from flask to an async
web framework (like aiohttp).  Then we would be guaranteed that access
to the dictionary would be single threaded, and would not need locks (as
long as we didnʼt await while looping through the dictionary, which
would not be necessary or even tempting to do).