Make the ZMQ interface use async
Previously, the zmq module used blocking calls.  This was fine as long
as it was the only thing running in the event loop, but for the stats
collection work I tried to put some other network I/O in the loop as
well, which was getting stalled.

Luckily, the python zmq bindings provide an async interface that we can
use without many changes.  Luckily², most of the methods in the zmq
library were already coroutines, even though they were not truly
asynchronous.  The only method which had to be changed to a coroutine
was zmq.Sender.send.

I tracked down (hopefully) all this methodʼs callers.  They came in
three places:
- the controller (Controller.report), which is called in a coroutine in
  __main__.py, where it can be awaited
- the runner, which eventually delegates to the Context class, whose
  public interface is through async context managers, where send can be
  awaited
- the stats sender (Stats.process), which is called by zmq.Receiver.run (as
  an entry in the listeners list).  This is not an async code path, so I
  opted to asyncio.ensure_future the send coroutine rather than awaiting
  it.