Rename the cdBuild.sh script (#58)
Because of the...interesting...implementation decision made by the core
platform tooling at
https://github.com/sky-uk/jenkins-tooling/pull/68/files#diff-23c13ce7b57a685655342611e7edeb84R35,
the presence of cdBuild.sh will block execution of the
defaultTaskExecution key.  If we want to have multiple pipelines, that
means that none of them can use the default cdBuild.sh name, because the
presence of this file will block all the other pipelines.  (Or more
accurately, all the defined pipelines will execute the cdBuild.sh file
no matter what).