Allow specifying custom buckets for mite http stats (#115)
For our chaos testing, weʼre interested in requests that take
longer (2-6 seconds).  The logarithmic buckets provide poor coverage for
this use case.  So weʼd like to override the buckets with something
different.  This lays the groundwork in mite that we need for that.

We use an environment variable for this, because the modular structure
of mite means that thereʼs not an easy way to pass a command line
argument in here.