Make the message bus configurable for the journey/scenario test commands
This allows us to do some fancy things.  Iʼve used this in chaos testing
to print out the request id of any request that takes longer than 5s.

(This change is technically not needed for the journey test command
itself, since we can print() from whereever we want in the code.  But it
allows us to test the message bus adapters that we write in the test
commands, eventually to plumb them into the message bus in a real
test – that is, this change facilitates debugging a custom component
like a stats, collector, etc. as part of journey test.)