Fix the http session pool memoization (#134)
Since the memoization was introduced, it had a defect which was that the
event loop was fetched at decorator execution (= function definition) time,
not at function call time.  We would not notice this in our core tests,
which all run under a single mite event loop, but it led to problems in the
id-mite-nft test suite that were importing and evaluating journey
functions in different event loops, importing outside loops, etc.

This fixes the memoization to occur at the correct time, and adds some
unit tests for the problem.  It also ports the contextmanager to use the
new (in 3.7) asynccontextmanager decorator for slightly less verbosity