Fix bug introduced with msgpack upgrade (#141)
strict_map_key prevents using integers as map keys when we unpack
messages (only str and bytes are allowed).  But the runner wire format
uses ints as keys.  The point of the setting of this option to true by
default is to be secure against hashtable-based DoS attacks.  But since
this is a mite-internal protocol, I think we are safe to set it to false.