Fix ci/cd jobs and how they run (#307)
What is the change?

Fixes to the pipelines. Because of how the env in which we run the jobs is setup, we can't fully customise the script names and the steps we run or their order. As a consequence we use the cdBuild script to select which is the right script to run based on the job name.